move VM tests from Blockchain.Test project to VM.Test project
