+++ src/Nethermind/Ethereum.VM.Test/Ethereum.VM.Test.csproj
@@ -31,7 +31,7 @@
       <Link>%(RecursiveDir)%(FileName)%(Extension)</Link>
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </EmbeddedResource>
-    <EmbeddedResource Include="..\..\tests\VMTests\**\*.*">
+    <EmbeddedResource Include="..\..\tests\GeneralStateTests\VMTests\**\*.*">
       <Link>%(RecursiveDir)%(FileName)%(Extension)</Link>
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </EmbeddedResource>
