+++ src/Nethermind/Ethereum.VM.Test/VmLogTests.cs
@@ -20,22 +20,18 @@ using System.Collections.Generic;
 using Ethereum.Test.Base;
 using NUnit.Framework;
 
-namespace Ethereum.Blockchain.Test
+namespace Ethereum.VM.Test
 {
     [TestFixture][Parallelizable(ParallelScope.All)]
-    // ReSharper disable once InconsistentNaming
-    public class VmIOAndFlowOperationsTests : GeneralStateTestBase
+    public class VmLogTests : GeneralStateTestBase
     {
         [TestCaseSource(nameof(LoadTests))]
         public void Test(GeneralStateTest test)
-        {
+        {    
             Assert.True(RunTest(test).Pass);
         }
         
-        public static IEnumerable<GeneralStateTest> LoadTests() 
-        { 
-            var loader = new TestsSourceLoader(new LoadGeneralStateTestsStrategy(), "vmIOAndFlowOperations");
-            return (IEnumerable<GeneralStateTest>)loader.LoadTests(); 
-        }
+        public static IEnumerable<GeneralStateTest> LoadTests() { var loader = new TestsSourceLoader(new LoadGeneralStateTestsStrategy(), "vmLogTest");
+            return (IEnumerable<GeneralStateTest>)loader.LoadTests(); }
     }
 }
