+++ src/Nethermind/Ethereum.VM.Test/VmPerformanceTests.cs
@@ -20,19 +20,20 @@ using System.Collections.Generic;
 using Ethereum.Test.Base;
 using NUnit.Framework;
 
-namespace Ethereum.Blockchain.Test
+namespace Ethereum.VM.Test
 {
     [TestFixture][Parallelizable(ParallelScope.All)]
     // ReSharper disable once InconsistentNaming
-    public class VmTests : GeneralStateTestBase
+    public class VmPerformanceTests : GeneralStateTestBase
     {
         [TestCaseSource(nameof(LoadTests))]
+        [Retry(3)]
         public void Test(GeneralStateTest test)
         {    
             Assert.True(RunTest(test).Pass);
         }
         
-        public static IEnumerable<GeneralStateTest> LoadTests() { var loader = new TestsSourceLoader(new LoadGeneralStateTestsStrategy(), "vmTests");
+        public static IEnumerable<GeneralStateTest> LoadTests() { var loader = new TestsSourceLoader(new LoadGeneralStateTestsStrategy(), "vmPerformance");
             return (IEnumerable<GeneralStateTest>)loader.LoadTests(); }
     }
 }
